Re-ubico y re-nombro ejemplo 13 a 31
