+++ Ejemplos/31/main.py
@@ -1,10 +1,13 @@
-from modules.timeutil import minutes, hours, days
-
 if __name__ == "__main__":
-    # Ejemplo "Transformar segundos a minutos, horas o dias" muestra el uso simple de Decorator para `type casting`
-    # un valor numerico decimal retorna por las operaciones de distintas funciones.
+    """
+    Ejemplo "Transformar segundos a minutos, horas o dias" muestra el uso simple de Decorator para `type casting`
+    un valor numerico decimal retorna por las operaciones de distintas funciones.
 
-    # Mas sobre Decorator https://docs.python.org/3/library/functools.html?highlight=wraps#functools.wraps
+    Mas sobre Decorator https://docs.python.org/3/library/functools.html?highlight=wraps#functools.wraps    
+    """
+
+    from timeutil import minutes, hours, days
+
     try:
         _seconds = int(input("Ingrese los segundos a transformar: "))
